Move ProgrammingLanguage from the android module to the shared module
